Changing DB initialisation to a no-op if permission and views doc already in place.
