+++ api/src/couchdb/helpers.ts
@@ -1,12 +1,17 @@
 /**
- * Gets a record, updates the _rev, then puts.
+ * Gets a record, updates the _rev, then puts or gracefully returns if
+ * writeOnClash is false.
  * @param db The database to upsert document into
  * @param data The document to replace
- * @returns The updated document (response from put)
+ * @param writeOnClash If the document already exists by ID, should it
+ * overwrite?
+ * @returns The updated document or document that was found if existing and
+ * writeOnClash = false
  */
-export async function upsertDocument<T extends {}>(
+export async function safeWriteDocument<T extends {}>(
   db: PouchDB.Database<T>,
-  data: PouchDB.Core.Document<T>
+  data: PouchDB.Core.Document<T>,
+  writeOnClash: boolean = true
 ) {
   try {
     // Try to get the existing document
@@ -18,6 +23,11 @@ export async function upsertDocument<T extends {}>(
       if (err.status !== 404) {
         throw err;
       }
+    }
+
+    if (existingDoc && !writeOnClash) {
+      // We don't overwrite an existing doc if writeOnClash is not true
+      return existingDoc;
     }
 
     // If the document exists, include its revision (which is overridden to be
