+++ api/src/couchdb/index.ts
@@ -19,30 +19,27 @@
  */
 
 import {
-  AuthDatabase,
-  AuthDatabaseSecurityDocument,
-  permissionDocument,
-  ProjectDataObject,
-  ProjectID,
-  ProjectMetaObject,
-  ProjectObject,
-  TemplateDetails,
-  viewsDocument,
+    AuthDatabase,
+    ProjectDataObject,
+    ProjectID,
+    ProjectMetaObject,
+    ProjectObject,
+    TemplateDetails,
 } from '@faims3/data-model';
 import PouchDB from 'pouchdb';
-import {initialiseJWTKey} from '../authkeys/initJWTKeys';
+import { initialiseJWTKey } from '../authkeys/initJWTKeys';
 import {
-  COUCHDB_INTERNAL_URL,
-  COUCHDB_PUBLIC_URL,
-  LOCAL_COUCHDB_AUTH,
+    COUCHDB_INTERNAL_URL,
+    COUCHDB_PUBLIC_URL,
+    LOCAL_COUCHDB_AUTH,
 } from '../buildconfig';
 import * as Exceptions from '../exceptions';
-import {upsertDocument} from './helpers';
 import {
-  initialiseDirectoryDB,
-  initialiseProjectsDB,
-  initialiseTemplatesDb,
-  initialiseUserDB,
+    initialiseAuthDb,
+    initialiseDirectoryDB,
+    initialiseProjectsDB,
+    initialiseTemplatesDb,
+    initialiseUserDB,
 } from './initialise';
 
 const DIRECTORY_DB_NAME = 'directory';
@@ -84,6 +81,21 @@ export const getDirectoryDB = (): PouchDB.Database | undefined => {
     }
   }
   return _directoryDB;
+};
+
+export const getAuthDB = (): AuthDatabase => {
+  if (!_authDB) {
+    const pouch_options = pouchOptions();
+    const dbName = COUCHDB_INTERNAL_URL + '/' + AUTH_DB_NAME;
+    try {
+      _authDB = new PouchDB(dbName, pouch_options);
+    } catch (error) {
+      throw new Exceptions.InternalSystemError(
+        'Error occurred while getting auth database.'
+      );
+    }
+  }
+  return _authDB;
 };
 
 /**
@@ -326,66 +338,5 @@ export const closeDatabases = async () => {
     } catch (error) {
       console.log(error);
     }
-  }
-};
-
-// ============
-// AUTH DB
-// ============
-
-export const getAuthDB = (): AuthDatabase => {
-  if (!_authDB) {
-    const pouch_options = pouchOptions();
-    const dbName = COUCHDB_INTERNAL_URL + '/' + AUTH_DB_NAME;
-    try {
-      _authDB = new PouchDB(dbName, pouch_options);
-    } catch (error) {
-      throw new Exceptions.InternalSystemError(
-        'Error occurred while getting auth database.'
-      );
-    }
-  }
-  return _authDB;
-};
-
-/**
- * Initialises the Auth DB by injecting a security document, and design
- * documents from the data model.
- *
- * NOTE: This is currently an upsert so that DBs can be maintained, but this
- * might be exploitable since the endpoint is not authenticated.
- *
- * @param db The database to initialise, intentionally not typed so as to not
- * include type errors for interacting with permission and security documents.
- */
-export const initialiseAuthDb = async (db: PouchDB.Database): Promise<void> => {
-  // can't save security on an in-memory database so skip if testing
-  if (process.env.NODE_ENV !== 'test') {
-    try {
-      await db.security(AuthDatabaseSecurityDocument).save();
-    } catch (e) {
-      console.error(
-        'Failed to save security document into Auth DB. Error: ' + e
-      );
-      throw e;
-    }
-  }
-
-  // Next setup the permissions document - overwrite if existing
-  try {
-    await upsertDocument(db, permissionDocument);
-  } catch (e) {
-    console.error(
-      'Failed to upsert permission document into Auth DB. Error: ' + e
-    );
-    throw e;
-  }
-
-  // Next setup the index document - overwrite if existing
-  try {
-    await upsertDocument(db, viewsDocument);
-  } catch (e) {
-    console.error('Failed to upsert index document into Auth DB. Error: ' + e);
-    throw e;
   }
 };
